third_party/blink/web_tests/platform/linux/virtual/new-shadowroot-attach-behavior-disabled/external/wpt/shadow-dom/declarative/declarative-shadow-dom-repeats-2-expected.txt

This is a testharness.js-based test.
[FAIL] Repeated declarative shadow roots keep only the first
  assert_equals: expected "1" but got "3"
Harness: the test ran to completion.
